+++ frontend/openchat-client/src/openchat.ts
@@ -42,6 +42,7 @@ import type {
     ChatEvent,
     ChatFrozenEvent,
     ChatIdentifier,
+    ChatListRoute,
     ChatListScope,
     ChatPermissions,
     ChatSummary,
@@ -55,6 +56,7 @@ import type {
     ClaimDailyChitResponse,
     ClientJoinCommunityResponse,
     ClientJoinGroupResponse,
+    CommunitiesRoute,
     CommunityDetailsResponse,
     CommunityIdentifier,
     CommunityInvite,
@@ -102,6 +104,7 @@ import type {
     GroupSearchResponse,
     GroupSubtype,
     HandleMagicLinkResponse,
+    HomeRoute,
     IdentityState,
     InternalBotCommandInstance,
     InviteCodeResponse,
@@ -150,6 +153,7 @@ import type {
     RemoveIdentityLinkResponse,
     RemoveMemberResponse,
     ResetInviteCodeResponse,
+    RouteParams,
     Rules,
     SaveCryptoAccountResponse,
     SearchDirectChatResponse,
@@ -161,6 +165,7 @@ import type {
     SetMemberDisplayNameResponse,
     SetPinNumberResponse,
     SetUsernameResponse,
+    ShareRoute,
     SiwePrepareLoginResponse,
     SiwsPrepareLoginResponse,
     SubmitProofOfUniquePersonhoodResponse,
@@ -280,33 +285,31 @@ import type { OpenChatConfig } from "./config";
 import { AIRDROP_BOT_USER_ID } from "./constants";
 import { snapshot } from "./snapshot.svelte";
 import {
-    app,
     askForNotificationPermission,
     bitcoinAddress,
     cryptoBalanceStore,
     cryptoLookup,
+    eventListScrollTop,
     exchangeRatesLookupStore,
     hasFlag,
     initNotificationStores,
     lastCryptoSent,
+    localUpdates,
+    mobileWidth,
     nervousSystemLookup,
+    notificationsSupported,
     notificationStatus,
+    rightPanelHistory,
     setSoftDisabled,
     swappableTokensStore,
-} from "./state/app.svelte";
+} from "./state";
+import { app } from "./state/app/state";
 import { botState } from "./state/bots.svelte";
-import { localUpdates } from "./state/global";
-import { pathState, type RouteParams } from "./state/path.svelte";
-import {
-    eventListScrollTop,
-    mobileWidth,
-    notificationsSupported,
-    rightPanelHistory,
-    ui,
-} from "./state/ui.svelte";
+import { pathState } from "./state/path/state";
+import { type RightPanelContent, ui } from "./state/ui/state";
 import type { UndoLocalUpdate } from "./state/undo";
 import { messagesRead, startMessagesReadTracker } from "./state/unread/markRead.svelte";
-import { userStore } from "./state/users/users.svelte";
+import { userStore } from "./state/users/state";
 import { diamondDurationToMs } from "./stores/diamond";
 import { applyTranslationCorrection } from "./stores/i18n";
 import { lastOnlineDates } from "./stores/lastOnlineDates";
@@ -2112,7 +2115,7 @@ export class OpenChat {
         localUpdates.markMessageDeleted(messageId, userId);
         undeletingMessagesStore.delete(messageId);
 
-        const recipients = [...app.selectedChat.userIds];
+        const recipients = [...app.selectedChatUserIds];
 
         rtcConnectionsManager.sendMessage(recipients, {
             kind: "remote_user_deleted_message",
@@ -2293,7 +2296,7 @@ export class OpenChat {
                 return false;
             });
 
-        this.#sendRtcMessage([...app.selectedChat.userIds], {
+        this.#sendRtcMessage([...app.selectedChatUserIds], {
             kind: "remote_user_toggled_reaction",
             id: chatId,
             messageId: messageId,
@@ -2458,7 +2461,7 @@ export class OpenChat {
             allUserIds.add(u);
         }
 
-        app.selectedChat.addUserIds([...allUserIds].filter((u) => u !== userId));
+        app.addSelectedChatUserIds([...allUserIds].filter((u) => u !== userId));
         await this.getMissingUsers(allUserIds);
     }
 
@@ -2748,7 +2751,7 @@ export class OpenChat {
 
         // this has the effect of clearing any state for any previously selected chat and
         // creating an empty container for the new chat's state
-        // app.setSelectedChat(chatId); TODO - make sure this works ok
+        app.setSelectedChat(chatId);
 
         const selectedChat = app.selectedChatSummary;
         if (selectedChat !== undefined) {
@@ -2785,7 +2788,7 @@ export class OpenChat {
         messageIndex: number,
         threadMessageIndex?: number,
     ): void {
-        const event = app.selectedChat.events.find(
+        const event = app.selectedChatEvents.find(
             (ev) => ev.event.kind === "message" && ev.event.messageIndex === messageIndex,
         ) as EventWrapper<Message> | undefined;
         if (event !== undefined) {
@@ -2852,7 +2855,7 @@ export class OpenChat {
                 makeRtcConnections(
                     app.currentUserId,
                     chat,
-                    app.selectedChat.threadEvents,
+                    app.selectedThreadEvents,
                     userStore.allUsers,
                     userStore.blockedUsers,
                     this.config.meteredApiKey,
@@ -2984,9 +2987,9 @@ export class OpenChat {
     }
 
     earliestLoadedThreadIndex(): number | undefined {
-        return app.selectedChat.threadEvents.length === 0
+        return app.selectedThreadEvents.length === 0
             ? undefined
-            : app.selectedChat.threadEvents[0].index;
+            : app.selectedThreadEvents[0].index;
     }
 
     previousThreadMessagesCriteria(thread: ThreadSummary): [number, boolean] {
@@ -3270,9 +3273,8 @@ export class OpenChat {
         updatedEvents: UpdatedEvent[],
     ): Promise<void> {
         const confirmedLoaded = confirmedEventIndexesLoaded(serverChat.id);
-        const confirmedThreadLoaded = app.selectedChat.confirmedThreadEventIndexesLoaded;
-        const selectedThreadRootMessageIndex =
-            app.selectedChat.selectedThread?.id?.threadRootMessageIndex;
+        const confirmedThreadLoaded = app.confirmedThreadEventIndexesLoaded;
+        const selectedThreadRootMessageIndex = app.selectedThreadId?.threadRootMessageIndex;
 
         // Partition the updated events into those that belong to the currently selected thread and those that don't
         const [currentChatEvents, currentThreadEvents] = updatedEvents.reduce(
@@ -3387,7 +3389,7 @@ export class OpenChat {
     }
 
     #confirmedThreadUpToEventIndex(): number | undefined {
-        const ranges = app.selectedChat.confirmedThreadEventIndexesLoaded.subranges();
+        const ranges = app.confirmedThreadEventIndexesLoaded.subranges();
         if (ranges.length > 0) {
             return ranges[0].high;
         }
@@ -3445,7 +3447,7 @@ export class OpenChat {
         threadRootMessageIndex: number | undefined,
     ): void {
         if (userId === app.currentUserId) {
-            const userIds = app.selectedChat.userIds;
+            const userIds = app.selectedChatUserIds;
             rtcConnectionsManager.sendMessage([...userIds], {
                 kind: "remote_user_removed_message",
                 id: chatId,
@@ -3563,8 +3565,7 @@ export class OpenChat {
                     mergeServerEvents(events, newEvents, context),
                 );
 
-                const selectedThreadRootMessageIndex =
-                    app.selectedChat.selectedThread?.id?.threadRootMessageIndex;
+                const selectedThreadRootMessageIndex = app.selectedThreadId?.threadRootMessageIndex;
                 if (selectedThreadRootMessageIndex !== undefined) {
                     const threadRootEvent = newEvents.find(
                         (e) =>
@@ -3600,7 +3601,7 @@ export class OpenChat {
         messageEvent: EventWrapper<Message>,
         threadRootMessageIndex: number | undefined,
     ): Promise<void> {
-        rtcConnectionsManager.sendMessage([...app.selectedChat.userIds], {
+        rtcConnectionsManager.sendMessage([...app.selectedChatUserIds], {
             kind: "remote_user_sent_message",
             id: clientChat.id,
             messageEvent: serialiseMessageForRtc(messageEvent),
@@ -3898,7 +3899,7 @@ export class OpenChat {
     }
 
     #rulesNeedAccepting(): boolean {
-        const chatRules = app.selectedChat.rules;
+        const chatRules = app.selectedChatRules;
         const chat = app.selectedChatSummary;
         if (chat === undefined || chatRules === undefined) {
             return false;
@@ -3943,8 +3944,8 @@ export class OpenChat {
     #eventsForMessageContext({
         threadRootMessageIndex,
     }: MessageContext): EventWrapper<ChatEvent>[] {
-        if (threadRootMessageIndex === undefined) return app.selectedChat.events;
-        return app.selectedChat.threadEvents;
+        if (threadRootMessageIndex === undefined) return app.selectedChatEvents;
+        return app.selectedThreadEvents;
     }
 
     eventExpiry(chat: ChatSummary, timestamp: number): number | undefined {
@@ -4124,7 +4125,9 @@ export class OpenChat {
     }
 
     markAllReadForCurrentScope() {
-        app.chatSummariesList.forEach((chat) => messagesRead.markAllRead(chat));
+        messagesRead.batchUpdate(() => {
+            app.chatSummaries.forEach((chat) => messagesRead.markAllRead(chat));
+        });
     }
 
     getDisplayDate = getDisplayDate;
@@ -4349,7 +4352,7 @@ export class OpenChat {
     }
 
     expandDeletedMessages(messageIndexes: Set<number>): void {
-        app.selectedChat.expandDeletedMessages(messageIndexes);
+        app.expandDeletedMessages(messageIndexes);
     }
 
     remoteUserToggledReaction(
@@ -4475,15 +4478,14 @@ export class OpenChat {
         if (
             selectedChat !== undefined &&
             chatIdentifiersEqual(fromChatId, selectedChat.id) &&
-            parsedMsg.threadRootMessageIndex ===
-                app.selectedChat.selectedThread?.id?.threadRootMessageIndex
+            parsedMsg.threadRootMessageIndex === app.selectedThreadId?.threadRootMessageIndex
         ) {
             this.#handleWebRtcMessageInternal(
                 fromChatId,
                 parsedMsg,
                 parsedMsg.threadRootMessageIndex === undefined
-                    ? app.selectedChat.events
-                    : app.selectedChat.threadEvents,
+                    ? app.selectedChatEvents
+                    : app.selectedThreadEvents,
                 parsedMsg.threadRootMessageIndex,
             );
         } else {
@@ -4593,7 +4595,7 @@ export class OpenChat {
     ): Promise<[UserSummary[], UserSummary[]]> {
         if (level === "channel") {
             // Put the existing channel members into a map for quick lookup
-            const channelMembers = newGroup ? undefined : app.selectedChat.members;
+            const channelMembers = newGroup ? undefined : app.selectedChatMembers;
 
             // First try searching the community members and return immediately if there are already enough matches
             // or if the caller does not have permission to invite users to the community
@@ -4640,7 +4642,7 @@ export class OpenChat {
                     const existing =
                         level === "community"
                             ? app.selectedCommunityMembers
-                            : app.selectedChat.members;
+                            : app.selectedChatMembers;
 
                     // Remove any existing members from the global matches until there are at most `maxResults`
                     // TODO: Ideally we would return the total number of matches from the server and use that
@@ -4660,7 +4662,7 @@ export class OpenChat {
         const communityMatches = this.#searchCommunityUsersForChannelInvite(
             searchTerm,
             maxResults,
-            app.selectedChat.members,
+            app.selectedChatMembers,
         );
 
         return Promise.resolve([communityMatches, []]);
@@ -5029,7 +5031,7 @@ export class OpenChat {
         const undo = localUpdates.updateChatMember(
             chatId,
             userId,
-            app.selectedChat.members.get(userId),
+            app.selectedChatMembers.get(userId),
             (m) => ({ ...m, role: newRole }),
         );
         return this.#sendRequest({ kind: "changeRole", chatId, userId, newRole })
@@ -5082,7 +5084,7 @@ export class OpenChat {
         // TODO get the list of exclusions from the user canister
 
         const exclusions = new Set<string>(
-            app.chatSummariesList
+            [...app.chatSummaries.values()]
                 .filter((c) => c.kind === "group_chat" && c.public)
                 .map((g) => chatIdentifierToString(g.id)),
         );
@@ -5813,7 +5815,7 @@ export class OpenChat {
             }
 
             // Update all users we have direct chats with
-            for (const chat of app.chatSummariesList) {
+            for (const chat of app.chatSummaries.values()) {
                 if (chat.kind == "direct_chat") {
                     usersToUpdate.add(chat.them.userId);
                 }
@@ -6708,7 +6710,7 @@ export class OpenChat {
     getUserLookupForMentions(): Record<string, UserOrUserGroup> {
         if (this.#userLookupForMentions === undefined) {
             const lookup = {} as Record<string, UserOrUserGroup>;
-            for (const [userId] of app.selectedChat.members) {
+            for (const [userId] of app.selectedChatMembers) {
                 let user = userStore.get(userId);
                 if (user !== undefined && app.selectedChatSummary?.kind === "channel") {
                     user = {
@@ -6934,7 +6936,7 @@ export class OpenChat {
             if (chat.kind === "direct_chat") {
                 userIds.push(chat.them.userId);
             } else if (this.isChatPrivate(chat)) {
-                userIds = [...app.selectedChat.members.keys()].filter((id) => id !== me);
+                userIds = [...app.selectedChatMembers.keys()].filter((id) => id !== me);
             }
             if (userIds.length > 0) {
                 await Promise.all(
@@ -7570,7 +7572,7 @@ export class OpenChat {
 
     selectFirstChat(): boolean {
         if (!get(mobileWidth)) {
-            const first = app.chatSummariesList.find((c) => !c.membership.archived);
+            const first = [...app.chatSummaries.values()].find((c) => !c.membership.archived);
             if (first !== undefined) {
                 pageRedirect(routeForChatIdentifier(app.chatListScope.kind, first.id));
                 return true;
@@ -7999,8 +8001,8 @@ export class OpenChat {
                             community: undefined,
                         };
 
-                        if (app.selectedChat.rules?.enabled ?? false) {
-                            acceptedRules.chat = app.selectedChat.rules?.version;
+                        if (app.selectedChatRules?.enabled ?? false) {
+                            acceptedRules.chat = app.selectedChatRules?.version;
                         }
 
                         if (app.selectedCommunityRules?.enabled ?? false) {
@@ -9213,6 +9215,47 @@ export class OpenChat {
                 }
             }
         }
+    }
+
+    pushRightPanelHistory(val: RightPanelContent) {
+        ui.pushRightPanelHistory(val);
+    }
+
+    popRightPanelHistory() {
+        ui.popRightPanelHistory();
+    }
+
+    filterRightPanelHistory(fn: (state: RightPanelContent) => boolean) {
+        return ui.filterRightPanelHistory(fn);
+    }
+
+    filterRightPanelHistoryByChatType(chat?: ChatSummary) {
+        return ui.filterRightPanelHistoryByChatType(chat);
+    }
+
+    isChatListRoute(route: RouteParams): route is ChatListRoute {
+        return pathState.isChatListRoute(route);
+    }
+
+    isHomeRoute(route: RouteParams): route is HomeRoute {
+        return pathState.isHomeRoute(route);
+    }
+
+    isCommunitiesRoute(route: RouteParams): route is CommunitiesRoute {
+        return pathState.isCommunitiesRoute(route);
+    }
+
+    isShareRoute(route: RouteParams): route is ShareRoute {
+        return pathState.isShareRoute(route);
+    }
+    setRouteParams(ctx: PageJS.Context, p: RouteParams) {
+        pathState.setRouteParams(ctx, p);
+    }
+    addUserGroupKey(key: string) {
+        app.addUserGroupKey(key);
+    }
+    toggleProposalFilterMessageExpansion(messageId: bigint, expand: boolean) {
+        app.toggleProposalFilterMessageExpansion(messageId, expand);
     }
 }
