--- frontend/openchat-client/src/state/chat/index.ts
@@ -1,1 +1,0 @@
-export { chatDetailsLocalUpdates } from "./detailsUpdates";
