--- frontend/openchat-client/src/state/community/index.ts
@@ -1,2 +1,0 @@
-export { communityLocalUpdates } from "./detailUpdates";
-export { CommunityDetailsState } from "./server";
