--- frontend/openchat-client/src/state/global/index.ts
@@ -1,1 +1,0 @@
-export { localUpdates } from "./local.svelte";
