+++ frontend/openchat-client/src/state/index.ts
@@ -1,8 +1,8 @@
-export * from "./app.svelte";
+export * from "./app/stores";
 export { FilteredProposals } from "./filteredProposals.svelte";
-export { localUpdates } from "./global/local.svelte";
-export * from "./path.svelte";
-export * from "./ui.svelte";
+export { localUpdates } from "./localUpdates";
+export * from "./path/stores";
+export * from "./ui/stores";
 export * from "./undo";
 export { messagesRead, type MessageReadState } from "./unread/markRead.svelte";
-export * from "./users/users.svelte";
+export * from "./users/stores";
