+++ frontend/openchat-client/src/state/unread/markRead.disabled.ts
@@ -1,7 +1,7 @@
 import { type EventWrapper, type GroupChatIdentifier, type Message } from "openchat-shared";
 import { SvelteMap } from "svelte/reactivity";
 import { vi } from "vitest";
-import { localUpdates } from "../global";
+import { localUpdates } from "../localUpdates";
 import { MessageReadTracker, MessagesRead } from "./markRead.svelte";
 
 const abcId: GroupChatIdentifier = { kind: "group_chat", groupId: "abc" };
